Implement a new EPUB renderer
We no longer use EPUB.js's Rendition, Manager, View, or other associated
classes. Instead, we have a new custom renderer that supports scrolled and
paginated views (the latter with some help from Readium CSS). It's much faster,
doesn't require offscreen views to be unloaded, and should be easier for us to
tweak and maintain in the future.

Some code is a bit messy and there are magic constants everywhere that need to
be factored out, so this should be considered a work in progress.

Caveats:
- Paginated view is not supported in Safari - its column layout is unusably slow
- No blank pages after sections end
- No support for pagination preferences embedded in the EPUB yet