Linting: Turn off prefer-const and stop using const in TypeScript code
When I first added TypeScript to this project, I set up ESLint with the
recommended config, which included prefer-const. I dutifully followed it,
thinking that that was this project's code style and not bothering to notice
that JS code wasn't using const nearly as heavily and that prefer-const had come
from my ESLint override for TypeScript, not from the root config. Once I did
notice, I procrastinated on turning it off and instead added more and more
const-heavy code. Oops!

From this point on, no more unnecessary const. It's awkward to type and almost
always pointless in TS code.