Remove DOMPurify completely
We rely on the CSP to prevent access to remote resources, and the CSP in
addition to the iframe sandbox (except in WebKit) to prevent scripts from
executing. DOMPurify is not ideal for our use case because it removes elements
from the tree, which could potentially change the EPUB CFIs and CSS selectors
that we use to target annotations.