EPUB/snapshot: Drag text when selection is under annotation
Ideally we would be able to disable pointer events on the annotation layer when
the pointer is over a selection, letting the drag event target the text instead
of the annotation, but doing that on pointerdown seems to be too late, and
something like a pointermove handler would be too heavy for an interaction this
rare. So we synthesize the drag image. A little weird, but I don't think many
people will notice.