aria and focus edits for vpat
- vpat 60: added role=button and aria-label to X clear
icon of search annotations
- vpat 61: hide magnifying glass of search annotations field
from screen readers
- vpat 62: explicitly refocus search annotations input field
on X click to not loose focus (possible with voiceover)
- vpat 64: added aria labels to annotation preview inputs in sidebar
- vpat 65: added title for context menu button
- vpat 69: aria-description for "tags" button.
It will always announce that this button can be clicked
to manage tags regardless what the current label is.
Note, there is a warning logged that aria-description
in not a valid aria property - it is a bug in react 17.
- vpat 71: properly announce annotations. Added listbox and option roles
to help screen readers handle cursor movements. aria-labelledby of annotation
points at header and aria-describedby points at the editable aria of
the annotation so that the announcement is "Page ..." followed by the content of
the annotation or the comment.
- vpat:72 add tab semantics to the sidebar component
vpat 72 states that we should have an audible announcement
after a new tab was selected. This does not seem exactly correct
per https://www.w3.org/WAI/WCAG21/Understanding/status-messages.html#excepted-examples,
however the sidebar should still have proper tabbar semantics
to help screen readers handle focus. That way, when a different
tab is selected, the screen readers tend to repeat the title
of the tab with a note that it is selected, so this should
address that concern.
- vpat 73: aria properties to announce thumbnails in the side bar
- vpat 74: added description and state to tags. Added aria-description
to tags to announce that this will filter annotations by
this tag. Note that there is a warning logged that aria-description
is not a valid aria property - it is a bug in react 17.
Added role=checkbox and aria-checked state to tags to indicate which ones are toggled on
- vpat 79: aria label for "Edit page number" input

- make annotation text focusable without editing
added tabstop parameter to ExpandableEditor that
will add data-tabstop=1 and tabindex=-1 if annotation is focused
regardless of readonly state. It allows us to focus the annotation
text via tabbing without breaking existing behavior of making
the annotation text focusable and editable only on double-click.

Fixes: zotero/zotero#4220
Addresses: zotero/zotero#4222
Addresses: zotero/zotero#4223