Add EPUB appearance customization popup
And add new styles for popups and sliders. Most slider styles are only
applied in Firefox/Zotero - Chrome/WebKit don't let you style the left
(filled) part of the slider separately, so trying to do in-depth
customization of the individual parts of the slider just makes it look
worse. We can set the accent color to get a little closer to the
intended look, though.