fix stuck focus and missing nvda announcement (#133)
- vpat 63: fix focus not being able to move forward
on tab from "Search annotations" input in the sidebar
if the search result is empty. Decide if the tabstop
should be added to annotationsView based on filteredAnnotations.length
 instead of unfiltered props.annotations.length
- vpat 73: fix NVDA not properly announcing the thumbnails
view when it is selected. The "Show thumbnails" label of the view
does not get announced, instead NVDA immediately announces page label
of the selected thumbnail, which is a bit confusing. This happens
because each time the selection changes, the view is re-added into the DOM,
which tricks NVDA into thinking that there is a new component that appeared
and it needs to be announced asap.
Intead, render outline, annotations and thumbnails views and just hide
views that are not selected. Also, added "Page" prefix to thumbnails
aria labels to give it a bit more context.
Followup to https://github.com/zotero/reader/pull/127