move virtual cursor when page is set in toolbar
So that screen reader users land on the right page
after changing the page number in the input or via
buttons next to it. Once the input is changed,
Escape keypress re-focuses the reader and it will
focus an element at the top of the page for virtual
cursor to move.