record top page to focus in _updateViewStats
- Update the target node to focus for in _updateViewStats
which triggers on every view scroll update. It allows
us to have one place where we record which page to
refocus, as opposed to doing it for every navigation
method
- To make sure our updates don't run too often and
are not cleared by the scroll listener right after
they are set, a11yWillPlaceVirtCursorOnTop runs
through a debounce which will set the node when
view updates are all finished.
- Removed timestamp of what the target was recorded
since it's irrelevant now
- Instead, added allowUpdates variable which will
not allow the node of _a11yVirtualCursorTarget to be
updated. It is needed for search results to not get
overriden by the a11yWillPlaceVirtCursorOnTop when
scrolling is done