move placeA11yVirtualCursor into utilities.js
So that views can share it without keeping it in
the reader.js.

Also, clear the virtual cursor target on mousedown
instead of during scroll - it also fixes the issue of
focus jumping to focused element on click after
scroll after an outline entry is selected. But it
also allows us to avoid multiple processes fighting
to set or clear virtual cursor and reduce debounce timeout
as a result.