vpat 81: instructions for keyboard interface
- added aria-descritions for buttons in the toolbar that
tell what shortcuts exist and how to use them for
different annotation types
- added aria-description to the input in 'Find in document'
popup to tell that one can create annotations from search
results
- when annotation is selected, announce a message
saying that the annotation is selected, indicating
that a popup appeared (if no sidebar), and telling
how one can manipulate the annotation (more/resize/etc.)

A few minor functional tweaks:
- when annotation is added from "Find in document",
close the popup and focus the view. Otherwise,
one is likely to close popup with Escape and then
they would have to search for the annotation again.
- when a note is added via keyboard shortcut,
focus the comment, since the instruction has to be
"Use ctrl-option-3 to add a note annotation", and
if comment is not focused, one would create an empty note
and then have to tab to the sidebar to edit it

Hardcoded color into the annotationManager as a temp
hack to finish up with a11y tweaks as there is regression
where keyboard shortcuts ctrl-option-1/2 from find
popup resuts in an error (post
ddfc5439304e47623185208de5e4573127054723).