navigate context menu by typing
- added keydown listener to contextmenu to select a menuitem by typing,
similar to native menus. After something is typed, find buttons with
text that begins with the input. If there is only one match, it is
clicked. If there are multiple, the first one is focused. The input
counter is reset after 2 seconds of not typing.
- listen to even candidates in focus-manager and dispatch copy of
keydown events from there to context-menu, in case one types when
the menu itself is not focused.

Also, fixed an encountered glitch where clicking inside of the rendered
reader content when the context menu is open would make the
annotation un-selectable via tab.