vpat 81: instructions for keyboard interface
- added aria-descritions for buttons in the toolbar that
tell what shortcuts exist and how to use them for
different annotation types
- added aria-description to the input in 'Find in document'
popup to tell that one can create annotations from search
results
- when annotation is selected, announce a message
saying that the annotation is selected, indicating
that a popup appeared (if no sidebar), and telling
how one can manipulate the annotation (more/resize/etc.)
- when annotation is added, announce that a new annotation
has been created. It helps in cases like when a text annotation
is added and text is immediately focused, or when underline
annotation is created from Find popup and focus
does not come back to the view to at least indicate
that it did work.
- when annotation is added from "Find in document",
close the popup.
- when a note is added via keyboard shortcut,
focus the comment, since the instruction has to be
"Use ctrl-option-3 to add a note annotation", and
if comment is not focused, one would create an empty note
and then have to tab to the sidebar to edit it
- fix glitchy announcements of external link annotations