miscellaneous a11y improvements
- announce focused annotations in epub and snapshot
similar to pdf
- announce citation when it is focused
- marking sidebar, toolbar and find popup with role="application"
so screen readers use focus mode instead of reading mode,
which is not appropriate to interact with clickable elements
- make outline values accessible to screen readers
- improved aria-live message announced during search navigation
to include the page number as well as the snippet of the
result if provided. It relies on currentSnippet and currentPageLabel
variables from FindState.
- added currentSnippet and currentPageLabel to find results
in EPUB and snapshots.
- add currentPageLabel to pdf-view findState. Need to have snippets
inclided as well
- added role="navigation" to start containers of epub ranges
so that screen readers indicate when one moves to a new page.
It also enabled navigation via d/shift-d for NVDA and r/shift-r
for JAWS to go to next/previous page as with PDFs.
Semanticaly, it is not correct but this seems to be the best
approach without adding nodes into the DOM, which will break CFIs.
- views have a new _a11yVirtualCursorTarget variable, in which
we record what node to make focusable and focus to help
screen readers place virtual cursor there. It is set to the following:
search match, the top of the page after scrolling (if not snapshot)
or the node matching the outline value (for snapshot). When the focus
returns into the content from elsewhere, _a11yVirtualCursorTarget will
become focusable. It is removed on pointer down events to avoid interference
with mouse interactions.

Note: JAWS looses virtual cursor if currently focused element is
removed or stop being focused, regardless where the cursor is.
For this reason, in PDF view, use the top-most page wrapper
node as cursor target, since anything inside of it is removed
when the page is unloaded.

into the page