Snapshot: Count sortIndex from <body>
Counting from <html> was silly because we mess with <head> pretty freely
(and annotations can never be outside the <body> anyway). Focus mode
forced this fix - the length of <head> changes when it's active, so sort
indexes can't be consistent unless we count from <body>.

This will likely require a server-side change to accept sortIndexes with
length 7.